ruby: Hide internals under miscreant/internals
Moves all internals to a special directory and adds :nodoc:

Some of these primitives are dangerous if used directly.
