+++ ruby/lib/miscreant/internals.rb
@@ -1,0 +1,12 @@
+# frozen_string_literal: true
+
+require "miscreant/internals/aes"
+require "miscreant/internals/aes/siv"
+require "miscreant/internals/aes/cmac"
+require "miscreant/internals/util"
+
+module Miscreant
+  # Internal functionality not intended for direct consumption
+  module Internals # :nodoc:
+  end
+end
