+++ ruby/spec/miscreant/internals/util_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe Miscreant::Util do
+RSpec.describe Miscreant::Internals::Util do
   describe "dbl" do
     let(:test_vectors) { described_class::DblExample.load_file }
